// frontend/src/app/core/models/Anexos/Anexo002_E28/Secciones.ts
export class Seccion1 {
    solicitante: string;
}
export class Seccion2 {
    servicio: string;
}

export class Seccion3 {
    localidad: string;
}

export class Seccion4 {
    finalidad: string;
    franjaEducativaComercial:string;
    franjaEducativaComunitaria:string;
}

export class Seccion5 {
    declaracion: boolean;
    porcentajeA: string;
    porcentajeB: string;
    porcentajeC: string;
    tipoDocumentoSolicitante: string;
    numeroDocumentoSolicitante: string;
    nombresSolicitante: string;
}